pyproject.toml
[project]
name = "curseforge-dl"
version = "0.1.0"
description = "CurseForge modpack downloader — download mods, resource packs, and shader packs from CurseForge."
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    { name = "kress" },
]
dependencies = [
    "httpx>=0.27",
    "pydantic>=2.0",
    "tqdm>=4.60",
    "click>=8.0",
    "python-dotenv>=1.0",
]

[project.scripts]
curseforge-dl = "curseforge_dl.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/curseforge_dl"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.24",
]
